requirements.txt
absl-py==2.0.0
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
astunparse==1.6.3
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.6.3
blinkt==0.1.2
buttonshim==0.0.2
cachetools==5.3.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
Cython==3.0.4
dbus-python==1.2.16
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.12.4
Flask==3.0.0
flatbuffers==20181003210633
fourletterphat==0.1.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.59.2
h5py==3.10.0
html5lib==1.1
idna==2.10
importlib-metadata==6.8.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
keras==2.14.0
lazy-object-proxy==0.0.0
libclang==16.0.6
logilab-common==1.8.1
lxml==4.6.3
Markdown==3.5.1
MarkupSafe==2.1.3
mccabe==0.6.1
microdotphat==0.2.1
ml-dtypes==0.2.0
mote==0.0.4
motephat==0.0.3
mpmath==1.3.0
mypy==0.812
mypy-extensions==0.4.3
networkx==3.2
numpy==1.26.1
oauthlib==3.1.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.2
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.12
picamerax==21.9.8
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==10.1.0
pkg_resources==0.0.0
protobuf==4.25.0
psutil==5.8.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.3.1
pycairo==1.16.2
pycups==2.0.1
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
pytesseract==0.3.10
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.8.1
pytz==2023.3.post1
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.6.0
simplejpeg==1.7.2
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-cpu-aws==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
textdistance==4.6.0
thonny==4.0.1
toml==0.10.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
touchphat==0.0.1
twython==3.8.2
typed-ast==1.4.2
typing_extensions==4.3.0
tzdata==2023.3
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.12.1
zipp==3.17.0
